# server/swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: FMD
  description: A website for movies' and TV shows' reviews
  termsOfService: http://swagger.io/terms/
  contact:
    email: om.dehlantube@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 3.0.0
servers:
- url: http://localhost:8080/VEDANTANEOGI_1/FMD_API3.0/3.0.0
...
paths:
...
  /review:
    post:
      tags:
      - Reviews
      summary: Add a new review
      operationId: review_post
      parameters:
      - name: movie_id
        in: query
        description: ID of the movie for which the review is being posted(null if
          TVShow)
        required: true
        style: form
        explode: false
        schema:
          type: string
          format: JSON
      - name: show_id
        in: query
        description: ID of the TVShow for which the review is being posted(null if
          Movie)
        required: true
        style: form
        explode: false
        schema:
          type: string
          format: JSON
      - name: user_id
        in: query
        description: ID of the user who is posting the review
        required: true
        style: form
        explode: false
        schema:
          type: integer
          example: 42
      - name: rating
        in: query
        description: rating by the user
        required: true
        style: form
        explode: false
        schema:
          maximum: 10
          minimum: 1
          type: integer
          example: 10
      - name: title
        in: query
        description: the review title of the review to be added
        required: true
        style: form
        explode: false
        schema:
          maxLength: 50
          type: string
          example: Amazing Watch!
      - name: content
        in: query
        description: content of the review
        required: true
        style: form
        explode: false
        schema:
          maxLength: 500
          type: string
          example: have you got colour in your cheeks? do you ever get that feel'
            that you can't shift the tide that sticks around like sumats in your teeth.
      - name: creation_time
        in: query
        description: creation time
        required: true
        style: form
        explode: false
        schema:
          type: string
          format: date-time
          example: 2023-04-02T12:30:45Z
      responses:
        "200":
          description: Review added successfully
        "400":
          description: Invalid request body
      x-openapi-router-controller: swagger_server.controllers.reviews_controller
...
  /review/movie/{id}:
    get:
      tags:
      - Reviews
      summary: Get all reviews of a movie
      description: Returns a list of reviews for a particular movie
      operationId: review_movie_id_get
      parameters:
      - name: id
        in: path
        description: ID of the Movie
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      - name: sort_type_reviews
        in: query
        description: Parameter based on which reviews will be sorted
        required: false
        style: form
        explode: false
        schema:
          type: string
          enum:
          - Likes
          - Recent
      - name: sort_order
        in: query
        description: sorting order
        required: false
        style: form
        explode: false
        schema:
          type: boolean
          enum:
          - true
          - false
      responses:
        "200":
          description: List of reviews for Movie
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/review'
                x-content-type: application/json
        "400":
          description: Invalid request
        "404":
          description: Title not found
      x-openapi-router-controller: swagger_server.controllers.reviews_controller
  /review/tvshow/{id}:
    get:
      tags:
      - Reviews
      summary: Get all reviews of a Tvshow
      description: Returns a list of reviews for a particular Tvshow
      operationId: review_tvshow_id_get
      parameters:
      - name: id
        in: path
        description: ID of the Tvshow
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      - name: sort_type_reviews
        in: query
        description: Parameter based on which reviews will be sorted
        required: false
        style: form
        explode: false
        schema:
          type: string
          enum:
          - Likes
          - Recent
      - name: sort_order
        in: query
        description: sorting order
        required: false
        style: form
        explode: false
        schema:
          type: boolean
          enum:
          - true
          - false
      responses:
        "200":
          description: List of reviews for TVShow
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/review'
                x-content-type: application/json
        "400":
          description: Invalid request
        "404":
          description: Title not found
      x-openapi-router-controller: swagger_server.controllers.reviews_controller
  /review/{review_id}/comment:
    get:
      tags:
      - Reviews
      summary: Get all comments of a review
      description: Returns a list of comments for a particular review
      operationId: review_review_id_comment_get
      parameters:
      - name: review_id
        in: path
        description: ID of the review
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 42
      responses:
        "200":
          description: List of comments for review
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/comment'
                x-content-type: application/json
        "400":
          description: Invalid request
        "404":
          description: Review not found
      x-openapi-router-controller: swagger_server.controllers.reviews_controller
    post:
      tags:
      - Reviews
      summary: Add comment to review
      operationId: review_review_id_comment_post
      parameters:
      - name: review_id
        in: path
        description: ID of the review
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 42
      - name: user_id
        in: query
        description: ID of the commentor
        required: true
        style: form
        explode: false
        schema:
          type: integer
          example: 42
      - name: content
        in: query
        description: Comment to be added to review
        required: true
        style: form
        explode: false
        schema:
          maxLength: 200
          type: string
          example: "im a puppet on a string, tracy island time travelling diamond,\
            \ coulda shaped heartaches"
      responses:
        "201":
          description: Comment added successfully to review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/comment'
        "400":
          description: Bad request. Invalid review ID or missing required fields
        "404":
          description: Review not found
      x-openapi-router-controller: swagger_server.controllers.reviews_controller
  /review/{review_id}/likes:
    put:
      tags:
      - Reviews
      summary: Like/Unlike Review
      operationId: review_review_id_likes_put
      parameters:
      - name: review_id
        in: path
        description: ID of the review
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 42
      - name: user_id
        in: query
        description: ID of the user who likes the review
        required: true
        style: form
        explode: false
        schema:
          type: integer
          example: 42
      responses:
        "201":
          description: Like added or removed
        "400":
          description: Bad request. Invalid review ID or missing required fields
        "404":
          description: Review not found
      x-openapi-router-controller: swagger_server.controllers.reviews_controller
...
  /review/user/{user_id}:
    get:
      tags:
      - Reviews
      summary: Get all reviews of a user
      description: Returns a list of reviews for a particular user
      operationId: review_user_user_id_get
      parameters:
      - name: user_id
        in: path
        description: ID of the user
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 42
      - name: sort_type_reviews
        in: query
        description: Parameter based on which reviews will be sorted
        required: false
        style: form
        explode: false
        schema:
          type: string
          enum:
          - Likes
          - Recent
      - name: sort_order
        in: query
        description: sorting order
        required: false
        style: form
        explode: false
        schema:
          type: boolean
          enum:
          - true
          - false
      responses:
        "200":
          description: List of reviews by the User
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/review'
                x-content-type: application/json
        "400":
          description: Invalid request
        "404":
          description: Title not found
      x-openapi-router-controller: swagger_server.controllers.reviews_controller
...
components:
  schemas:
...
    movie_id:
      type: object
      properties:
        id:
          type: integer
          example: 42
      example:
        id: 42
    show_id:
      type: object
      properties:
        id:
          type: integer
          example: 42
      example:
        id: 42
...
    review:
      type: object
      properties:
        review_id:
          type: integer
          example: 42
        title:
          maxLength: 50
          type: string
          example: Amazing Watch!
        movie_id:
          $ref: '#/components/schemas/movie_id'
        show_id:
          $ref: '#/components/schemas/show_id'
        user_id:
          type: integer
          example: 42
        likes:
          type: object
          additionalProperties:
            type: boolean
          example:
            "1": true
            "2": false
            "42": true
        rating:
          maximum: 10
          minimum: 1
          type: integer
          example: 10
        content:
          maxLength: 500
          type: string
          example: have you got colour in your cheeks? do you ever get that feel'
            that you can't shift the tide that sticks around like sumats in your teeth.
        creation_time:
          type: string
          format: date-time
          example: 2023-04-02T12:30:45Z
      example:
        creation_time: 2023-04-02T12:30:45Z
        review_id: 42
        show_id:
          id: 42
        user_id: 42
        rating: 10
        title: Amazing Watch!
        movie_id:
          id: 42
        content: have you got colour in your cheeks? do you ever get that feel' that
          you can't shift the tide that sticks around like sumats in your teeth.
        likes:
          "1": true
          "2": false
          "42": true
    comment:
      type: object
      properties:
        comment_id:
          type: integer
          example: 42
        review_id:
          type: integer
          example: 42
        user_id:
          type: integer
          example: 42
        content:
          maxLength: 200
          type: string
          example: "im a puppet on a string, tracy island time travelling diamond,\
            \ coulda shaped heartaches"
      example:
        review_id: 42
        user_id: 42
        comment_id: 42
        content: "im a puppet on a string, tracy island time travelling diamond, coulda\
          \ shaped heartaches"
...